Half step

+++ Server/default_configure.h
@@ -8,7 +8,7 @@ namespace default_configures
 
 std::string IP = "127.0.0.1";
 
-uint32_t Port = 15000;
+uint16_t Port = 15000;
 
 std::string Directory = "./";
